fix service node ovs failure
Service node bridges (br-int) were failing to start up due to
potentially interfering configuration in /etc/openvswitch/conf.db.
Added commands to start.sh to stop OVS, remove its conf, then restart.

Also modified "prepare sdk" script to create backup copies of important
files and replace them with "current" files from docker-devstack.
--> Need to rebuild containers.

+++ service/build/prepare_sdk.sh
@@ -7,27 +7,27 @@ sudo -E apt-get install exuberant-ctags
 cd $HOMEDIR || exit
 
 # set up docker-devstack (development)
-[ ! -d 'docker-devstack/.git' ] && git clone https://github.com/matt-welch/docker-devstack
-source ${HOMEDIR}/devstack/openrc/admin demo
-printenv | grep OS_ > openstackrc
-sed -i 's:^OS_:export OS_:' openstackrc
-cp openstackrc ${DEV_DIR}/openstack/
+[ ! -d "${HOMEDIR}/docker-devstack/.git" ] && git clone https://github.com/matt-welch/docker-devstack ${HOMEDIR}/docker-devstack
+
+if [ -f ${HOMEDIR}/devstack/stacking.status ]; then
+    source ${HOMEDIR}/devstack/openrc/admin demo
+    printenv | grep OS_ > openstackrc
+    sed -i 's:^OS_:export OS_:' openstackrc
+    cp openstackrc ${DEV_DIR}/openstack/
+fi
 
 # move files and replace with newer copies
 mkdir -p $BACK
-FILE='start.sh'
-OLDFILE=${HOMEDIR}/$FILE
-NEWFILE=${DEV_DIR}/$FILE
-[ -f "$OLDFILE" ] && mv "$OLDFILE" $BACK
-ln -s $NEWFILE $HOMEDIR/
-FILE='restart.sh'
-OLDFILE=${HOMEDIR}/$FILE
-NEWFILE=${DEV_DIR}/$FILE
-[ -f "$OLDFILE" ] && mv "$OLDFILE" $BACK
-ln -s $NEWFILE $HOMEDIR/
+for FILE in start.sh restart.sh service.odl.local.conf; do
+    OLDFILE=${HOMEDIR}/$FILE
+    NEWFILE=${DEV_DIR}/$FILE
+    [ -f "$OLDFILE" ] && mv "$OLDFILE" $BACK
+    echo "Linking $NEWFILE in $HOMEDIR"
+    ln -s $NEWFILE $HOMEDIR/
+done
 
 # setup daffy-dotfiles (terminal settings)
-[ ! -d 'daffy-dotfiles/.git' ] && git clone https://github.com/matt-welch/daffy-dotfiles
+[ ! -d "${HOMEDIR}/daffy-dotfiles/.git" ] && git clone https://github.com/matt-welch/daffy-dotfiles ${HOMEDIR}/daffy-dotfiles
 cd ${HOMEDIR}/daffy-dotfiles || exit
 ${HOMEDIR}/daffy-dotfiles/install.sh
